[Chore] feedback (#46)
