--- ACON-iOS/ACON-iOS/Network/Image/Withdrawal/DTO/WithdrawalRequest.swift
@@ -1,15 +1,0 @@
-//
-//  WithdrawalRequest.swift
-//  ACON-iOS
-//
-//  Created by Jaehyun Ahn on 2/18/25.
-//
-
-import Foundation
-
-struct WithdrawalRequest: Codable {
-    
-    let reason : String
-    let refreshToken : String
-
-}
