--- ACON-iOS/ACON-iOS/Network/Image/Withdrawal/WithdrawalTargetType.swift
@@ -1,45 +1,0 @@
-//
-//  WithdrawalTargetType.swift
-//  ACON-iOS
-//
-//  Created by Jaehyun Ahn on 2/18/25.
-//
-
-import Foundation
-
-import Moya
-
-enum WithdrawalTargetType {
-    
-    case postWithdrawal(requestBody: WithdrawalRequest)
-    
-}
-
-extension WithdrawalTargetType: TargetType {
-    
-    var path: String {
-        switch self {
-        case .postWithdrawal:
-            return utilPath + "members/withdrawal"
-        }
-    }
-    
-    var method: Moya.Method {
-        switch self {
-        case .postWithdrawal:
-            return .post
-        }
-    }
-    
-    var task: Task {
-        switch self {
-        case .postWithdrawal(let data):
-            return .requestJSONEncodable(data)
-        }
-    }
-    
-    var headers: [String: String]? {
-        let headers = HeaderType.headerWithToken()
-        return headers
-    }
-}
